requirements.txt
Flask==2.1.0
google-cloud-datastore==2.15.1
Flask-OAuthlib
requests
flask_bcrypt
json
uuid
ast